improvements
* doc strings according to PEP
* skipped tests marked properly
* refactored CLI
