+++ nichtparasoup/cli/__init__.py
@@ -1,0 +1,4 @@
+"""Subpackage containing all of nichtparasoup's command line interface related code
+"""
+
+# This file intentionally does not import submodules
