+++ nichtparasoup/cli/commands.py
@@ -1,13 +1,7 @@
 import logging
-from typing import Any, List, Optional
+from typing import Any, Optional
 
 from nichtparasoup._internals import _message, _message_exception
-
-
-def _logging_init(level: int) -> None:  # pragma: no cover
-    if not logging.root.handlers:
-        logging.root.setLevel(level)
-        logging.root.addHandler(logging.StreamHandler())
 
 
 class Commands(object):
@@ -15,10 +9,13 @@ class Commands(object):
     @staticmethod
     def run(config_file: Optional[str] = None) -> int:
         from os.path import abspath
+
+        from nichtparasoup._internals import _logging_init
         from nichtparasoup.config import get_config, get_imagecrawler
         from nichtparasoup.core import NPCore
         from nichtparasoup.core.server import Server as ImageServer
         from nichtparasoup.webserver import WebServer
+
         try:
             config = get_config(abspath(config_file) if config_file else None)
             _logging_init(getattr(logging, config['logging']['level']))
@@ -44,10 +41,12 @@ class Commands(object):
     @staticmethod
     def config_dump_file(config_file: str) -> int:
         from os.path import abspath, isfile
+
+        from nichtparasoup._internals import _confirm
         from nichtparasoup.config import dump_defaults
+
         config_file = abspath(config_file)
         if isfile(config_file):
-            from nichtparasoup._internals import _confirm
             overwrite = _confirm('File already exists, overwrite?')
             if overwrite is not True:
                 _message('Abort.')
@@ -62,7 +61,9 @@ class Commands(object):
     @staticmethod
     def config_check_file(config_file: str) -> int:
         from os.path import abspath
+
         from nichtparasoup.testing.config import ConfigFileTest
+
         config_file = abspath(config_file)
         config_test = ConfigFileTest()
         try:
@@ -80,21 +81,23 @@ class Commands(object):
             _message_exception(ValueError('exactly one action required'))
             return 255
         action, action_value = active_actions.popitem()
-        return dict(
+        return dict(  # type: ignore
             version=cls.info_version,
             imagecrawler_list=cls.info_imagecrawler_list,
             imagecrawler_desc=cls.info_imagecrawler_desc,
         )[action](action_value)
 
     @staticmethod
-    def info_version(_: Any) -> int:
+    def info_version(_: Optional[Any] = None) -> int:
         from nichtparasoup import VERSION
+
         _message(VERSION)
         return 0
 
     @staticmethod
-    def info_imagecrawler_list(_: Any) -> int:
+    def info_imagecrawler_list(_: Optional[Any] = None) -> int:
         from nichtparasoup.imagecrawler import get_imagecrawlers
+
         imagecrawlers = get_imagecrawlers().names()
         if not imagecrawlers:
             _message_exception(Warning('no ImageCrawler found'))
@@ -107,6 +110,7 @@ class Commands(object):
         from nichtparasoup._internals import _log
         from nichtparasoup.core.server import type_module_name_str
         from nichtparasoup.imagecrawler import get_imagecrawlers
+
         imagecrawler_class = get_imagecrawlers().get_class(imagecrawler)
         if not imagecrawler_class:
             _message_exception(ValueError('unknown ImageCrawler {!r}'.format(imagecrawler)))
@@ -132,13 +136,3 @@ class Commands(object):
                 'Class: {!r}'.format(type_module_name_str(imagecrawler_class)),
             ]))
         return 0
-
-
-def main(args: Optional[List[str]] = None) -> int:
-    from nichtparasoup.cmdline.argparse import parser as argparser
-    options = dict(argparser.parse_args(args=args).__dict__)
-    if options.pop('debug', False):
-        _logging_init(logging.DEBUG)
-        _message('DEBUG ENABLED :)', 'cyan')
-    command = options.pop('command')
-    return getattr(Commands, command)(**options)  # type: ignore
