+++ nichtparasoup/testing/imagecrawler.py
@@ -103,8 +103,7 @@ class ImageCrawlerLoaderTest(TestCase, ABC):
     @property
     @abstractmethod
     def ic_class(self) -> Type[BaseImageCrawler]:  # pragma: no cover
-        """
-        the class of your ImageCrawler.
+        """the class of your ImageCrawler.
 
         Example implementation:
             return MyImageCrawler
