+++ tests/test_10_nichtparasoup/test_cli/test_main.py
@@ -1,0 +1,14 @@
+import unittest
+
+from nichtparasoup.cli.main import main as main_
+
+
+class CommandsRunTest(unittest.TestCase):
+    """Current implementation is just a preparation for later.
+    """
+
+    @unittest.skip("nothing to test, yet")
+    def test_nothing(self) -> None:
+        """Current implementation is a placeholder, so the imported main_ stays active.
+        """
+        main_()
