--- tests/test_10_nichtparasoup/test_cmdline/test_commands.py
@@ -1,49 +1,0 @@
-import unittest
-
-
-class CommandsRunTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsConfigTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsConfigDumpFileTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsConfigCheckFileTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsInfoTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsInfoVersionTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsInfoImagecrawlerListTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
-
-
-class CommandsInfoImagecrawlerDescTest(unittest.TestCase):
-
-    def test_(self) -> None:
-        self.skipTest('TODO')
