--- tests/test_10_nichtparasoup/test_cmdline/test_main.py
@@ -1,7 +1,0 @@
-import unittest
-
-
-class MainTest(unittest.TestCase):
-
-    def test_main(self) -> None:
-        self.skipTest('TODO')
